Move ontology elements to type package
