+++ metadata-transformer/src/main/java/no/unit/nva/metadata/service/MetadataService.java
@@ -23,11 +23,11 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.stream.Collectors;
 import no.unit.nva.api.CreatePublicationRequest;
-import no.unit.nva.metadata.Bibo;
-import no.unit.nva.metadata.Citation;
-import no.unit.nva.metadata.DcTerms;
+import no.unit.nva.metadata.type.Bibo;
+import no.unit.nva.metadata.type.Citation;
+import no.unit.nva.metadata.type.DcTerms;
 import no.unit.nva.metadata.MetadataConverter;
-import no.unit.nva.metadata.OntologyProperty;
+import no.unit.nva.metadata.type.OntologyProperty;
 import org.apache.any23.extractor.ExtractionException;
 import org.eclipse.rdf4j.model.IRI;
 import org.eclipse.rdf4j.model.Model;
