+++ metadata-transformer/src/main/java/no/unit/nva/metadata/type/OntologyProperty.java
@@ -1,4 +1,4 @@
-package no.unit.nva.metadata;
+package no.unit.nva.metadata.type;
 
 import org.eclipse.rdf4j.model.IRI;
 import org.eclipse.rdf4j.model.ValueFactory;
