+++ metadata-transformer/src/test/java/no/unit/nva/metadata/service/MetadataServiceTest.java
@@ -3,8 +3,8 @@ package no.unit.nva.metadata.service;
 import com.github.tomakehurst.wiremock.WireMockServer;
 import j2html.tags.EmptyTag;
 import no.unit.nva.api.CreatePublicationRequest;
-import no.unit.nva.metadata.Citation;
-import no.unit.nva.metadata.DcTerms;
+import no.unit.nva.metadata.type.Citation;
+import no.unit.nva.metadata.type.DcTerms;
 import no.unit.nva.metadata.service.testdata.ContributorArgumentsProvider;
 import no.unit.nva.metadata.service.testdata.DcContentCaseArgumentsProvider;
 import no.unit.nva.metadata.service.testdata.LanguageArgumentsProvider;
